mekhane/anamnesis/requirements.txt
# Gnōsis v0.1 Requirements

# Core
arxiv>=2.1.0
requests>=2.31.0
PyYAML>=6.0

# Vector DB
lancedb>=0.4.0

# Embedding
onnxruntime>=1.16.0
tokenizers>=0.15.0
numpy>=1.24.0

# Optional (for stats)
pandas>=2.0.0
